backend/requirements.txt
click==8.1.3
dnspython==2.2.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
flask-sock==0.5.2
geographiclib==1.52
geopy==2.2.0
gunicorn==20.1.0
h11==0.14.0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pymongo==4.3.2
python-dotenv==0.21.0
simple-websocket==0.8.1
six==1.16.0
Turbo-Flask==0.8.3
Werkzeug==2.2.2
wsproto==1.2.0
zipp==3.9.0
